fixed the wierdness of double incrementing and crashing when removing a word from the active set in study view.
When making changes to the active set, the TagPeer subscribe or remove must be called BEFORE addToCurrentSet of removeFromCurrentSet.

Right now this happens in 3 places. We should make a manager or something that everyone talks to to add or remove, and it determines if it's the current set or not.